refactor: separete price oracle inputs into new files
